backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
motor==3.5.0
pymongo==4.8.0
pydantic[email]==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyMuPDF==1.24.0
openai==1.50.0
python-multipart==0.0.9
httpx==0.27.0
pytest==8.3.0
pytest-asyncio==0.24.0
